requirements.txt
aiohttp==3.11.12
asyncio==3.4.3
bs4==0.0.2
fastapi==0.115.8
httpx==0.28.1
lxml_html_clean==0.4.1
python-dotenv==1.0.1
pytest==8.3.4
readability-lxml==0.8.1
redis==5.2.1
requests==2.32.3
uvicorn==0.34.0
